refactor: Optimize pipe handling to reduce CPU usage and prevent duplicates

+++ lib/discourse_systray/systray.rb
@@ -638,22 +638,19 @@ module ::DiscourseSystray
         # Create pipe if it doesn't exist
         system("mkfifo #{PIPE_PATH}") unless File.exist?(PIPE_PATH)
         
-        loop do
-          begin
-            # Open pipe in read-only mode
-            File.open(PIPE_PATH, "r") do |pipe|
-              while line = pipe.gets
-                puts line
-              end
+        # Create pipe if it doesn't exist
+        system("mkfifo #{PIPE_PATH}") unless File.exist?(PIPE_PATH)
+        
+        # Single continuous read from pipe
+        begin
+          File.open(PIPE_PATH, "r") do |pipe|
+            while line = pipe.gets
+              puts line
+              STDOUT.flush
             end
-          rescue Errno::ENOENT, IOError => e
-            # If pipe is deleted or parent process exits, we should exit too
-            puts "Pipe closed or deleted, exiting: #{e}" if OPTIONS[:debug]
-            break
-          rescue StandardError => e
-            puts "Error reading pipe: #{e}" if OPTIONS[:debug]
-            sleep 0.1  # Avoid tight loop if there are temporary errors
           end
+        rescue Errno::ENOENT, IOError => e
+          puts "Pipe closed or deleted, exiting: #{e}" if OPTIONS[:debug]
         end
       else
         return if self.class.running?
@@ -673,15 +670,10 @@ module ::DiscourseSystray
     end
 
     def publish_to_pipe(msg)
-      # Create pipe if it doesn't exist
-      system("mkfifo #{PIPE_PATH}") unless File.exist?(PIPE_PATH)
-      
-      # Open pipe in non-blocking mode to prevent hanging
+      return unless File.exist?(PIPE_PATH)
       begin
-        Timeout.timeout(0.1) do
-          File.open(PIPE_PATH, "w") { |f| f.puts(msg) }
-        end
-      rescue Timeout::Error, Errno::EPIPE, IOError => e
+        File.open(PIPE_PATH, "w") { |f| f.puts(msg) }
+      rescue Errno::EPIPE, IOError => e
         puts "Error writing to pipe: #{e}" if OPTIONS[:debug]
       end
     end
